Simplify: dedupe formatTimeSlot, memoize grouping, extract cap-height class
- Move the duplicated formatTimeSlot helper out of PollResults.tsx and
  PollPageClient.tsx into lib/timeUtils, reusing formatDayLabel and
  formatDurationLabel instead of re-implementing them.
- Extract the inline text-box-trim style into a .cap-height-text CSS
  class in globals.css so both the results grid and the voting ballot
  share one rule (and drop the React.CSSProperties cast).
- Memoize slotsByDay in TimeResults to match TimeSlotBubbles, and drop
  the now-unnecessary [...options].sort() — slot keys already arrive
  in chronological order from the backend.
- Trim narrating JSX comments that restated Tailwind classes.
