+++ components/TimeSlotBubbles.tsx
@@ -64,13 +64,11 @@ export default function TimeSlotBubbles({
         const { weekday, monthDay } = formatStackedDayLabel(dateStr);
         return (
         <div key={dateStr} className="flex gap-2 items-start py-3 first:pt-0 last:pb-0">
-          {/* Day label — stacked, left-aligned, narrow fixed width for consistent bubble alignment */}
           <div className="w-12 shrink-0 pt-1 text-xs font-medium text-gray-500 dark:text-gray-400 text-left leading-tight">
             <div>{weekday}</div>
             <div>{monthDay}</div>
           </div>
 
-          {/* Bubbles — fixed width so labels sit in vertically aligned columns */}
           <div className="flex flex-wrap gap-1.5">
             {slots.map((slot, idx) => {
               const state = getState(slot);
@@ -99,16 +97,7 @@ export default function TimeSlotBubbles({
                       : "bg-white dark:bg-gray-800 border-gray-300 dark:border-gray-600 text-gray-700 dark:text-gray-300 hover:border-gray-400 dark:hover:border-gray-400 focus:ring-blue-400",
                   ].join(" ")}
                 >
-                  <span
-                    className="block"
-                    style={{
-                      lineHeight: 1,
-                      textBoxTrim: 'trim-both',
-                      textBoxEdge: 'cap alphabetic',
-                    } as React.CSSProperties}
-                  >
-                    {label}
-                  </span>
+                  <span className="block cap-height-text">{label}</span>
                   {excluded > 0 && (
                     <span className="absolute -top-1.5 -right-1.5 flex h-4 w-4 items-center justify-center rounded-full bg-orange-500 text-[9px] font-bold text-white leading-none pointer-events-none">
                       {excluded}
